Start moving RGB backlight back to Notification service
