+++ applications/services/application.fam
@@ -12,6 +12,5 @@ App(
         "loader",
         "power",
         "namechanger_srv",
-        "rgb_backlight",
     ],
 )
