+++ applications/services/notification/application.fam
@@ -4,7 +4,7 @@ App(
     apptype=FlipperAppType.SERVICE,
     entry_point="notification_srv",
     cdefines=["SRV_NOTIFICATION"],
-    requires=["input","rgb_backlight"],
+    requires=["input"],
     provides=["notification_settings"],
     stack_size=int(1.5 * 1024),
     order=100,
