--- applications/services/rgb_backlight/application.fam
@@ -1,10 +1,0 @@
-App(
-    appid="rgb_backlight",
-    name="RgbBackLightSrv",
-    apptype=FlipperAppType.SERVICE,
-    entry_point="rgb_backlight_srv",
-    cdefines=["SRV_RGB_BACKLIGHT"],
-    stack_size=1 * 1024,
-    order=1290,
-    sdk_headers=["rgb_backlight.h"],
-)
