+++ targets/f7/api_symbols.csv
@@ -1,5 +1,5 @@
 entry,status,name,type,params
-Version,+,85.0,,
+Version,+,86.0,,
 Header,+,applications/drivers/subghz/cc1101_ext/cc1101_ext_interconnect.h,,
 Header,+,applications/services/bt/bt_service/bt.h,,
 Header,+,applications/services/bt/bt_service/bt_keys_storage.h,,
@@ -38,7 +38,6 @@ Header,+,applications/services/locale/locale.h,,
 Header,+,applications/services/notification/notification.h,,
 Header,+,applications/services/notification/notification_messages.h,,
 Header,+,applications/services/power/power_service/power.h,,
-Header,+,applications/services/rgb_backlight/rgb_backlight.h,,
 Header,+,applications/services/rpc/rpc_app.h,,
 Header,+,applications/services/storage/storage.h,,
 Header,+,lib/bit_lib/bit_lib.h,,
@@ -415,10 +414,6 @@ Function,-,LL_mDelay,void,uint32_t
 Function,-,Osal_MemCmp,int,"const void*, const void*, unsigned int"
 Function,-,Osal_MemCpy,void*,"void*, const void*, unsigned int"
 Function,-,Osal_MemSet,void*,"void*, int, unsigned int"
-Function,+,SK6805_get_led_count,uint8_t,
-Function,+,SK6805_init,void,
-Function,+,SK6805_set_led_color,void,"uint8_t, uint8_t, uint8_t, uint8_t"
-Function,+,SK6805_update,void,
 Function,-,SystemCoreClockUpdate,void,
 Function,-,SystemInit,void,
 Function,-,_Exit,void,int
@@ -3138,9 +3133,6 @@ Function,-,putw,int,"int, FILE*"
 Function,-,qsort,void,"void*, size_t, size_t, __compar_fn_t"
 Function,-,qsort_r,void,"void*, size_t, size_t, int (*)(const void*, const void*, void*), void*"
 Function,-,quick_exit,void,int
-Function,+,rainbow_timer_start,void,RGBBacklightApp*
-Function,+,rainbow_timer_starter,void,RGBBacklightApp*
-Function,+,rainbow_timer_stop,void,RGBBacklightApp*
 Function,+,rand,int,
 Function,-,rand_r,int,unsigned*
 Function,+,random,long,
@@ -3159,12 +3151,6 @@ Function,-,remquol,long double,"long double, long double, int*"
 Function,-,rename,int,"const char*, const char*"
 Function,-,renameat,int,"int, const char*, int, const char*"
 Function,-,rewind,void,FILE*
-Function,+,rgb_backlight_get_color_count,uint8_t,
-Function,+,rgb_backlight_get_color_text,const char*,uint8_t
-Function,+,rgb_backlight_set_led_static_color,void,"uint8_t, uint8_t"
-Function,+,rgb_backlight_settings_load,void,RGBBacklightSettings*
-Function,+,rgb_backlight_settings_save,void,const RGBBacklightSettings*
-Function,+,rgb_backlight_update,void,float
 Function,-,rindex,char*,"const char*, int"
 Function,-,rint,double,double
 Function,-,rintf,float,float
